Removing utilAPIs.go and moving surviving code into containerd.go
